[#57] feat: Restructure modal page
