+++ src/components/organisms/main/status.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { View, Image } from "react-native";
 import { Aim, Point, RecordsOfMonth } from "../../molecules/index";
-import SetGoalModal from "../../../pages/setGoalModal/setGoalModal";
+import SetGoalModal from "../../../pages/modals/setGoalModal/setGoalModal";
 import styles from "./status.styles";
 
 const Status = ({
