+++ src/pages/modals/errorModal/errorModal.styles.ts
