+++ src/pages/modals/setGoalModal/setGoalModal.style.ts
